Link to the latest API reference docs for each SDK in the sidebar.
We previously had an issue where _any_ release would become listed as the latest
release for an SDK if it had a version number greater than anything previous,
even if it was a pre-release (alpha, beta, etc).

This has since been fixed and I've manually updated the `releases.json` file to
have correct entries for the latest release for each SDK.

This means we can now revert to using the `latest` entry in `releases.json` for
each SDK to populate the API reference docs link in the sidebar rather than
harcoding 3.0.1, which is no longer the latest (we're up to 3.0.6).